KIP-415: Refactor ExtendedWorkerState as a separate public class
